# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bike Trip Journal API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/trips/{slug}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get a trip's metadata and its bikes",
        "description": "**Context.** A trip from before accounts was shared through two links, a rider\nlink and a viewer link. Since decision-log Entry 29 either link only *locates*\nthe trip; what the caller may do is decided by their membership. This is the\nendpoint both links open first. It answers three questions in one round trip \u2014\nwhich trip is this, which bikes are on it, and should the UI offer writes \u2014\nbecause the app is used on the Stuart Hwy where a second request is a second\nchance to be offline. Tasks `t-trip-metadata-endpoint`, `t-am-write-gate-legacy`.\n\n**How it works.** `{slug}` is resolved by the shared `require_trip_access`\ndependency, which accepts **either** slug and raises a 404 if neither matches.\nThe session is optional here and never produces a 401. The dependency derives\n`access`: `rider` iff the session user has an active membership on this trip\n(read fresh on every request), `viewer` for everyone else \u2014 anonymous, a\nsigned-in non-member, a pending requester or a revoked member. Which slug was\nfollowed plays no part. `viewer.role` is derived alongside it (`anonymous`,\n`none`, `pending`, `rider` or `leader`), and `access` is `rider` exactly when\nthat role is `rider` or `leader`. `visibility`, `publicDelayHours`, `riderCount`\nand `lastPublicStopAt` are filled as on `GET /api/v2/trips/{tripId}`; this read\nitself stays full and undelayed for either slug. The trip's bikes are fetched by `data/repositories/bikes.list_by_trip`,\nfiltered to this trip; a trip with no bikes returns `\"bikes\": []`, which is a\nnormal trip and not a 404. Neither slug is in the response: `TripOut` has no\nslug field, and the slug is the credential.\n\n`access` is a **UI hint**, deprecated, and not the enforcement point. It tells\nthe frontend whether to render Add stop / upload photo / edit bikes. The write\nendpoints check the session and membership themselves regardless of what the UI\nchose to show.\n\n**Related APIs.** `GET /api/trips/{slug}/stops` for the stops this header sits\nabove, `GET /api/trips/{slug}/map` for the same trip as GeoJSON, and\n`POST /api/trips/{slug}/bikes` / `PATCH /api/trips/{slug}/bikes/{id}` for the\nmember-only writes behind the `bikes` list returned here.",
        "operationId": "get_trip_api_trips__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's rider or viewer slug \u2014 the link the trip was shared with. Either one locates the trip; neither grants anything. Legacy reads accept both and give the full, undelayed trip.",
              "title": "Slug"
            },
            "description": "The trip's rider or viewer slug \u2014 the link the trip was shared with. Either one locates the trip; neither grants anything. Legacy reads accept both and give the full, undelayed trip."
          }
        ],
        "responses": {
          "200": {
            "description": "The trip, the bikes on it, and whether the caller is an active member.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this slug. Deliberately the same answer for a mistyped link, a revoked one and a guess \u2014 see `docs/api-contract.md`, 'Access control: 403 and 404 are different answers'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "A path parameter failed validation. This operation takes no request body, so the path is the only thing that can fail; its path parameters are plain strings, so in practice it does not return this today.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `public-read` limit: 120 requests a minute per client address. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/me/trips": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "The caller's trips",
        "description": "**Context.** \"Your trips\": where a signed-in rider finds the trips they belong\nto, private ones included (decision-log Entry 29).\n\n**How it works.** Gated by `require_session` (401 without a valid session).\nLists the trips the caller has an **active** membership on, with `role`\n`rider` or `leader`. Revoked and self-departed memberships are left out, and a\npending join request is not a membership (see `GET /api/v2/me/join-requests`).\nNewest membership first, then by trip `id`. Not paged. An account with no trips\ngets `200` with `[]`.\n\n**Related APIs.** `GET /api/v2/trips/{tripId}` opens one; `POST /api/v2/trips`\ncreates one; `GET /api/v2/trips` is the public list.",
        "operationId": "list_my_trips_api_v2_me_trips_get",
        "responses": {
          "200": {
            "description": "Every trip the caller is an active member of, with their role.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MyTripOut"
                  },
                  "type": "array",
                  "title": "Response List My Trips Api V2 Me Trips Get"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: none sent, expired, signed out or revoked, or the account is disabled. If a session cookie was sent it is cleared (`Max-Age=0`). Carries `WWW-Authenticate: Cookie realm=\"bike-trip-journal\"`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `public-read` limit: 120 requests a minute per client address. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1.0
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/trips": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "List public trips",
        "description": "**Context.** The Discover list: the public trips anyone can browse, with no\naccount (decision-log Entry 29). Private trips never appear.\n\n**How it works.** Anonymous: a session, if sent, is ignored, and the response is\nthe same for every caller. Each item carries `riderCount` (active members,\nleaders included) and `lastPublicStopAt`, the latest `arrivedAt` among stops\nvisible to the public \u2014 those with `arrivedAt <= now() - publicDelayHours` on the\nserver clock. Sorted by `lastPublicStopAt` descending with nulls last, then `id`\nascending; trips with no visible stop come last. Pages hold `limit` items\n(default 20, 1-50). `nextCursor` is opaque; pass it back as `cursor` unchanged\nfor the next page. It is `null` on the last page. A cursor this endpoint did not\nissue is a `422`.\n\n**Related APIs.** `GET /api/v2/trips/{tripId}` for one trip,\n`GET /api/v2/me/trips` for the trips the caller is a member of.",
        "operationId": "list_public_trips_api_v2_trips_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The `nextCursor` from the previous page, unchanged. Omit it for the first page. Anything this endpoint did not issue is rejected with 422 / VALIDATION_ERROR.",
              "title": "Cursor"
            },
            "description": "The `nextCursor` from the previous page, unchanged. Omit it for the first page. Anything this endpoint did not issue is rejected with 422 / VALIDATION_ERROR."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "How many trips per page, from 1 to 50 inclusive; defaults to 20. Anything else is rejected with 422 / VALIDATION_ERROR.",
              "default": 20,
              "title": "Limit"
            },
            "description": "How many trips per page, from 1 to 50 inclusive; defaults to 20. Anything else is rejected with 422 / VALIDATION_ERROR."
          }
        ],
        "responses": {
          "200": {
            "description": "One page of public trips, most recently active first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPageOut"
                }
              }
            }
          },
          "422": {
            "description": "`limit` is outside 1-50, or `cursor` is not a `nextCursor` this endpoint issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `public-read` limit: 120 requests a minute per client address. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Create a trip",
        "description": "**Context.** How a signed-in rider starts a new trip (decision-log Entry 29). The\ncaller becomes its first leader. Trips created here have no slugs: they are\nreached by id.\n\n**How it works.** `limit_trip_create` runs first (3 a day per account; a replay\nspends nothing), then `require_session` (401). `TripCreate.id` is\nclient-generated and must be a canonical lowercase UUID (422 otherwise). It then\ndecides one of three branches: an unseen id creates the trip -> `201`; an id the\ncaller created and is still an active member of is a replay -> `200` with the\nstored trip; any other existing id -> `409`, nothing disclosed. A caller who has\nalready created 20 trips gets `409` for a new id. In one transaction the trip is\nstored with `visibility` as sent (default `public`), `publicDelayHours` 24 and\nno slugs, together with the caller's `leader` membership. Parallel creates by one\naccount are serialised, so the cap can't be overshot.\n\n**Related APIs.** `GET /api/v2/me/trips` lists the caller's trips;\n`PATCH /api/v2/trips/{tripId}` edits the settings; `GET /api/v2/trips/{tripId}`\nreads it back.",
        "operationId": "create_trip_api_v2_trips_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The trip just created, with the caller as its leader (201).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "200": {
            "description": "**A replay, not a second trip.** The caller created a trip with this id and is still an active member of it, so nothing was created and the **stored** trip is returned, even where this request's body differs from it. `viewer.role` is the caller's current role on it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: none sent, expired, signed out or revoked, or the account is disabled. If a session cookie was sent it is cleared (`Max-Age=0`). The offline queue pauses on this and resumes after sign-in. Nothing was written.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Either the `id` already names a trip that is not the caller's own replay -- another account's trip, a trip from before accounts, or one the caller created but has since left or been removed from -- or the caller has already created 20 trips (the lifetime cap). Nothing was created, and nothing about another trip is disclosed. Never-retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The body failed validation: `id` is not a canonical lowercase UUID, `name` is empty or over 100 characters after trimming, `startDate` is not an ISO date, or `visibility` is not 'public'/'private'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `trip-create` limit: 3 creates a day per account, or per client address for a request with no session. A replay of the caller's own trip spends no token and is never refused by it. Checked before the session is validated, so nothing was written. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/trips/{tripId}": {
      "patch": {
        "tags": [
          "trips"
        ],
        "summary": "Update a trip's settings",
        "description": "**Context.** A leader's trip settings: its name, whether it is public, and how\nlong stops stay hidden from the public (decision-log Entry 29).\n\n**How it works.** `limit_writes` runs first, then `require_trip_leader`: a valid\nsession (401, for every trip id alike), then the trip (404 if it doesn't exist,\nor if it is private and the caller has no membership row on it -- the same 404\nthe v2 reads give), then an active **leader** membership (403). Only the fields\npresent in the body change; omitting a field is how \"no change\" is spelled, an\nexplicit `null` is a 422, and an empty body is a 200 with nothing changed.\nThere is no server-side publish confirmation. Last write wins.\n\n**Related APIs.** `GET /api/v2/trips/{tripId}` reads the settings back;\n`POST /api/v2/trips` creates the trip.",
        "operationId": "patch_trip_api_v2_trips__tripId__patch",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's id. Not a secret: it only names the trip. Leader actions need a signed-in account with an active leader membership on it. No session is a 401 for every trip id, existing or not; a private trip the caller has no membership row on is a 404 identical to a trip id that doesn't exist; anyone else who is not an active leader -- a non-member of a public trip, a revoked member, an active rider -- is a 403.",
              "title": "Tripid"
            },
            "description": "The trip's id. Not a secret: it only names the trip. Leader actions need a signed-in account with an active leader membership on it. No session is a 401 for every trip id, existing or not; a private trip the caller has no membership row on is a 404 identical to a trip id that doesn't exist; anyone else who is not an active leader -- a non-member of a public trip, a revoked member, an active rider -- is a 403."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The trip after applying the patch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: none sent, expired, signed out or revoked, or the account is disabled. Checked **before** the trip is looked up, so the answer is the same for a public trip, a private trip and a trip id that doesn't exist. If a session cookie was sent it is cleared (`Max-Age=0`). Nothing was changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, and the trip is visible to the caller (public, or they have a membership row on it), but they are not an active leader: \"You're not a rider on this trip.\" for a non-member of a public trip, \"You're no longer a rider on this trip.\" for a revoked member, \"You're not a leader on this trip.\" for an active rider. Nothing was changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this id, **or** the trip is private and the caller has no membership row on it. The two are byte-identical, as on the v2 reads, so a trip id can't be used to learn whether a private trip exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The body failed validation: an explicit `null`, a `name` empty or over 100 characters after trimming, a `visibility` other than 'public'/'private', or a `publicDelayHours` outside 0-168.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `writes` limit: 600 requests an hour per account, or per client address for a request with no session. Checked before the session, the trip and the membership, so nothing was changed. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get a trip",
        "description": "**Context.** A trip's page, located by its id rather than a slug (decision-log\nEntry 29). Anyone can open a public trip; a private one is visible to its\nactive members only.\n\n**How it works.** `{tripId}` is resolved by `require_trip_reader`: the trip must exist and be\npublic, or the caller must be an active member (rider or leader); anything else\nis a 404 identical to a trip id that doesn't exist. The session is optional and\nnever produces a 401. Active members see everything with no delay. Anyone else\nsees only stops with `arrivedAt <= now() - publicDelayHours`, on the server\nclock at request time. `viewer.role` is computed from the session on every\nrequest: `anonymous`, `none`, `pending`, `rider` or `leader`. `access` is\ndeprecated and is `rider` exactly when the role is `rider` or `leader`.\n`riderCount` and `lastPublicStopAt` are the same for every caller, members\nincluded; `lastPublicStopAt` is the latest `arrivedAt` among stops visible to\nthe public. Bikes are not delayed. No slug, username or user id is in the\nresponse.\n\n**Related APIs.** `GET /api/v2/trips/{tripId}/stops`, `/map`, `/bikes` and\n`/stops/{stopId}/photos` for the rest of the trip; `GET /api/v2/trips` for the\npublic list.",
        "operationId": "get_trip_api_v2_trips__tripId__get",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's id. Not a secret: it only names the trip. A private trip the caller may not see answers exactly as a trip id that doesn't exist does.",
              "title": "Tripid"
            },
            "description": "The trip's id. Not a secret: it only names the trip. A private trip the caller may not see answers exactly as a trip id that doesn't exist does."
          }
        ],
        "responses": {
          "200": {
            "description": "The trip, its bikes, and what the caller is to it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this id, **or** the trip is private and the caller is not an active member. The two are byte-identical -- status, body and headers -- so a trip id can't be used to learn whether a private trip exists. Never a 401: a read needs no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "A path parameter failed validation. This operation takes no request body, so the path is the only thing that can fail; its path parameters are plain strings, so in practice it does not return this today.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `public-read` limit: 120 requests a minute per client address. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/me/join-requests": {
      "get": {
        "tags": [
          "join-requests"
        ],
        "summary": "The caller's join requests",
        "description": "**Context.** Where a requester sees how their requests stand (decision-log\nEntry 29): waiting, approved, not approved or cancelled.\n\n**How it works.** `limit_public_read` runs first, then `require_session` (401).\nLists every request the caller has made, in any state, newest `createdAt` first,\nat most 100, each with its trip's name. A request a leader blocked is reported\nas `rejected`, and no cooldown date is given. An account with no requests gets\n`200` with `[]`.\n\n**Related APIs.** `POST /api/v2/trips/{tripId}/join-requests` makes one;\n`POST /api/v2/join-requests/{requestId}/cancel` withdraws one;\n`GET /api/v2/me/trips` lists the trips the caller is already on.",
        "operationId": "list_my_join_requests_api_v2_me_join_requests_get",
        "responses": {
          "200": {
            "description": "The caller's join requests, newest first, at most 100.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MyJoinRequestOut"
                  },
                  "type": "array",
                  "title": "Response List My Join Requests Api V2 Me Join Requests Get"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: none sent, expired, signed out or revoked, or the account is disabled. If a session cookie was sent it is cleared (`Max-Age=0`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `public-read` limit: 120 requests a minute per client address. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1.0
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TripOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The trip's id. Stable, and safe to show or log \u2014 it is not a credential and cannot be used to reach the trip. Neither slug appears anywhere in this response."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The trip's display name, as the rider titled it."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "title": "Startdate",
            "description": "The day the trip starts, as an ISO 8601 date (YYYY-MM-DD). A date with no time and no timezone: it is the calendar day the journal opens on, not an instant, so it reads the same wherever the viewer is."
          },
          "bikes": {
            "items": {
              "$ref": "#/components/schemas/BikeOut"
            },
            "type": "array",
            "title": "Bikes",
            "description": "The bikes on **this** trip, embedded so the app's first load is one request rather than two. Empty list when no bikes have been added yet \u2014 that is a normal trip, not an error. Order is not part of the contract; match bikes by `id`."
          },
          "access": {
            "$ref": "#/components/schemas/Access",
            "description": "Deprecated; kept for clients from before accounts. 'rider' when the caller is signed in with an active membership on this trip (rider or leader), 'viewer' otherwise \u2014 anonymous, a signed-in non-member, a pending requester or a revoked member. Which slug was used plays no part. A UI hint for whether to show write UI (Add stop, upload photo, edit bikes); the server enforces access on every write regardless."
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility",
            "description": "Who can read the trip: 'public' (anyone, with the public delay applied for non-members) or 'private' (members only; anyone else gets the same 404 as for a trip that doesn't exist)."
          },
          "publicDelayHours": {
            "type": "integer",
            "title": "Publicdelayhours",
            "description": "How many hours a stop stays hidden from non-members after its `arrivedAt`, from 0 to 168. Members see every stop at once, whatever this says."
          },
          "riderCount": {
            "type": "integer",
            "title": "Ridercount",
            "description": "How many people are on the trip: its active members, leaders included. A count only; who they are is never in this response."
          },
          "lastPublicStopAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastpublicstopat",
            "description": "The latest `arrivedAt` among the stops visible to the public, i.e. with the public delay applied, as a timezone-aware ISO 8601 instant. The same value for every caller, members included. Null when no stop is visible to the public yet."
          },
          "viewer": {
            "$ref": "#/components/schemas/ViewerOut",
            "description": "What the caller is to this trip, computed from the session on every request. Replaces `access`. A UI hint only: the server enforces access on every write regardless."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "startDate",
          "bikes",
          "access",
          "visibility",
          "publicDelayHours",
          "riderCount",
          "lastPublicStopAt",
          "viewer"
        ],
        "title": "TripOut",
        "description": "One trip: `GET /api/trips/{slug}` (either slug) and `GET /api/v2/trips/{tripId}`.\n\nCarries no slug, no username and no user id -- a non-member can receive it."
      },
...
    }
  }
}